+++ server/ui/routes.go
@@ -68,7 +68,7 @@ func (c *UIController) RegisterRoutes(app *echo.Echo) {
 				_ = ctx.String(http.StatusInternalServerError, "Something went very wrong!")
 			}
 		}(ctx)
-		requestedPath := ctx.Request().URL.Path
+		requestedPath := path.Clean(ctx.Request().URL.Path)
 
 		// Even though we are using an embedded filesystem for the UI, we still want
 		// to make sure we do not use relative paths.
